backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfminer.six
python-docx
spacy
google-generativeai
python-dotenv
pandas
openpyxl
pydantic
gunicorn